Tests/LibWeb/Text/expected/Editing/clipboard-shortcut-copy-cut.txt

clipboard after copy: null
editor text after copy: "hello world"
clipboard after cut: null
editor text after cut: "hello world"
clipboard after canceled copy: null
